# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API KoriPay",
        "description": "Interface interactive pour tester les endpoints de la plateforme financière KoriPay (Projet L3 IAGE / ISI Dakar).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Serveur Local de Développement"
        }
    ],
    "paths": {
...
        "/admin/depot": {
            "post": {
                "tags": [
                    "Module Admin : Opérations Guichet"
                ],
                "summary": "Effectuer un dépôt de fonds sur le compte d'un client",
                "description": "Permet à un administrateur ou un agent authentifié de créditer le compte d'un client depuis la console de guichet React. Génère une écriture comptable et transmet une facture par e-mail au bénéficiaire.",
                "operationId": "adminDepotFonds",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "montant"
                                ],
                                "properties": {
                                    "telephone": {
                                        "description": "Numéro de téléphone du client bénéficiaire",
                                        "type": "string",
                                        "example": "+221771234567"
                                    },
                                    "montant": {
                                        "description": "Montant en FCFA à déposer",
                                        "type": "number",
                                        "example": 50000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dépôt effectué avec succès et solde mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statut": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dépôt effectué avec succés !"
                                        },
                                        "donnees": {
                                            "description": "Détails du reçu de dépôt généré",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Opération rejetée car le compte client cible est suspendu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statut": {
                                            "type": "string",
                                            "example": "erreur"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Impossible de créditer ce compte car il est actuellement suspendu."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Action non autorisée (l'utilisateur connecté n'est ni admin ni agent)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statut": {
                                            "type": "string",
                                            "example": "erreur"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Action non autorisée."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun client KoriPay actif trouvé avec ce numéro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statut": {
                                            "type": "string",
                                            "example": "erreur"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aucun client trouvé avec ce numéro de téléphone"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation (champs obligatoires manquants ou montant inférieur au minimum)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statut": {
                                            "type": "string",
                                            "example": "erreur"
                                        },
                                        "erreurs": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Incident technique ou crash système au cours de l'écriture en base de données",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statut": {
                                            "type": "string",
                                            "example": "erreur"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Une erreur technique est survenue lors du dépôt."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/retrait/initier": {
            "post": {
                "tags": [
                    "Module Admin : Opérations Guichet"
                ],
                "summary": "Étape 1 : Initialiser un retrait au guichet",
                "description": "Permet à un administrateur ou un agent de guichet de lancer une demande de retrait pour un client. Un code OTP à 6 chiffres est généré et envoyé par e-mail au client.",
                "operationId": "adminInitierRetrait",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "montant"
                                ],
                                "properties": {
                                    "telephone": {
                                        "description": "Téléphone du client qui retire",
                                        "type": "string",
                                        "example": "+221771234567"
                                    },
                                    "montant": {
                                        "description": "Montant du retrait",
                                        "type": "number",
                                        "example": 25000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code OTP généré et envoyé par e-mail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statut": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Code OTP de validation généré et envoyé par e-mail au client."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solde insuffisant pour effectuer le retrait"
                    },
                    "403": {
                        "description": "Action non autorisée (réservé aux admins et agents)"
                    },
                    "404": {
                        "description": "Aucun client trouvé avec ce numéro"
                    },
                    "422": {
                        "description": "Erreur de validation des données fournies"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez votre jeton de sécurité (Token Sanctum) reçu lors du Login pour tester les routes protégées.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}